Restructure commons.core -> commons
