+++ src/main/java/com/notably/commons/Version.java
@@ -1,4 +1,4 @@
-package com.notably.commons.core;
+package com.notably.commons;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
