+++ src/main/java/com/notably/commons/path/AbsolutePath.java
@@ -1,10 +1,10 @@
-package com.notably.commons.core.path;
+package com.notably.commons.path;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import com.notably.commons.core.path.exceptions.InvalidPathException;
+import com.notably.commons.path.exceptions.InvalidPathException;
 
 /**
  * Represents the Path to a Block, starting from the Root node.
