+++ src/main/java/com/notably/commons/path/Path.java
@@ -1,4 +1,4 @@
-package com.notably.commons.core.path;
+package com.notably.commons.path;
 
 import java.util.List;
