+++ src/main/java/com/notably/commons/path/exceptions/InvalidPathException.java
@@ -1,6 +1,6 @@
-package com.notably.commons.core.path.exceptions;
+package com.notably.commons.path.exceptions;
 
-import com.notably.commons.core.path.Path;
+import com.notably.commons.path.Path;
 
 /**
  * {@link RuntimeException} to be thrown when an invalid representation of a path is supplied during
