+++ src/main/java/com/notably/logic/LogicManager.java
@@ -4,8 +4,8 @@ import java.nio.file.Path;
 import java.util.List;
 import java.util.logging.Logger;
 
-import com.notably.commons.core.GuiSettings;
-import com.notably.commons.core.LogsCenter;
+import com.notably.commons.GuiSettings;
+import com.notably.commons.LogsCenter;
 import com.notably.logic.commands.Command;
 import com.notably.logic.commands.exceptions.CommandException;
 import com.notably.logic.parser.NotablyParser;
