+++ src/main/java/com/notably/logic/commands/OpenCommand.java
@@ -2,7 +2,7 @@ package com.notably.logic.commands;
 
 import static java.util.Objects.requireNonNull;
 
-import com.notably.commons.core.path.AbsolutePath;
+import com.notably.commons.path.AbsolutePath;
 import com.notably.model.Model;
 
 /**
