+++ src/main/java/com/notably/model/Model.java
@@ -2,7 +2,7 @@ package com.notably.model;
 
 import java.nio.file.Path;
 
-import com.notably.commons.core.GuiSettings;
+import com.notably.commons.GuiSettings;
 import com.notably.model.block.BlockModel;
 import com.notably.model.suggestion.SuggestionModel;
 import com.notably.model.viewstate.ViewStateModel;
