+++ src/main/java/com/notably/model/ReadOnlyUserPrefs.java
@@ -2,7 +2,7 @@ package com.notably.model;
 
 import java.nio.file.Path;
 
-import com.notably.commons.core.GuiSettings;
+import com.notably.commons.GuiSettings;
 
 /**
  * Unmodifiable view of user prefs.
