+++ src/main/java/com/notably/model/block/BlockTree.java
@@ -1,6 +1,6 @@
 package com.notably.model.block;
 
-import com.notably.commons.core.path.AbsolutePath;
+import com.notably.commons.path.AbsolutePath;
 
 /**
  * API of BlockTree component.
