+++ src/test/java/com/notably/logic/suggestion/commands/DeleteSuggestionCommandTest.java
@@ -10,8 +10,8 @@ import java.util.Optional;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
-import com.notably.commons.core.path.AbsolutePath;
-import com.notably.commons.core.path.exceptions.InvalidPathException;
+import com.notably.commons.path.AbsolutePath;
+import com.notably.commons.path.exceptions.InvalidPathException;
 import com.notably.model.Model;
 import com.notably.model.ModelManager;
 import com.notably.model.block.Block;
