+++ src/test/java/com/notably/storage/JsonUserPrefsStorageTest.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import com.notably.commons.core.GuiSettings;
+import com.notably.commons.GuiSettings;
 import com.notably.commons.exceptions.DataConversionException;
 import com.notably.model.UserPrefs;
