+++ src/test/java/com/notably/storage/StorageManagerTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import com.notably.commons.core.GuiSettings;
+import com.notably.commons.GuiSettings;
 import com.notably.model.UserPrefs;
 
 public class StorageManagerTest {
